feat: restore ALWAYS_ALLOW_PRIORITY and disable autoAddToPolicyByDefault
