+++ packages/cli/src/config/settingsSchema.ts
@@ -1501,7 +1501,7 @@ const SETTINGS_SCHEMA = {
         label: 'Auto-add to Policy by Default',
         category: 'Security',
         requiresRestart: false,
-        default: true,
+        default: false,
         description: oneLine`
           When enabled, the "Allow for all future sessions" option becomes the
           default choice for low-risk tools in trusted workspaces.
