+++ packages/core/src/policy/persistence.test.ts
@@ -6,8 +6,10 @@
 
 import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
 import * as path from 'node:path';
-import { createPolicyUpdater } from './config.js';
-import { getAlwaysAllowPriorityFraction } from './utils.js';
+import {
+  createPolicyUpdater,
+  getAlwaysAllowPriorityFraction,
+} from './config.js';
 import { PolicyEngine } from './policy-engine.js';
 import { MessageBus } from '../confirmation-bus/message-bus.js';
 import { MessageBusType } from '../confirmation-bus/types.js';
