+++ packages/core/src/policy/utils.ts
@@ -5,21 +5,6 @@
  */
 
 /**
- * Priority used for user-defined "Always allow" rules.
- * This is above extension rules but below user-defined TOML rules.
- */
-export const ALWAYS_ALLOW_PRIORITY = 3.95;
-
-/**
- * Returns the fractional priority of ALWAYS_ALLOW_PRIORITY scaled to 1000.
- */
-export function getAlwaysAllowPriorityFraction(): number {
-  return Math.round(
-    (ALWAYS_ALLOW_PRIORITY - Math.floor(ALWAYS_ALLOW_PRIORITY)) * 1000,
-  );
-}
-
-/**
  * Escapes a string for use in a regular expression.
  */
 export function escapeRegex(text: string): string {
